+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 
 * Fully removes references to the A2 `self.partial` module method. It appeared only once outside of comments, but was not actually used by the UI. The `self.render` method should be used for simple template rendering.
 
+### Changes
+
+* Adds localization keys for the password field component's min and max error messages.
+
 ## 3.8.1 - 2021-11-23
 
 ### Fixes